bird: add validate_cmd for bird.conf
We should validate the bird configuration before Puppet actually applies
the change with validate_cmd (calling bird to do the validation),
helping prevent situations in which the configuration file was incorrect
and the bird service failed.

This is not a perfect solution -- ideally, we should have something in
the CI itself but that's for later (ha!).

This was inspired by the deployment of bird2 (T310574).

Hosts: doh1001.wikimedia.org, dns1001.wikimedia.org
Change-Id: Iac5f64f9606dfd14b19ef450262a08be08553976