labweb: point tlsproxy envoy at port 8080 for striker
Add hiera config for local tlsproxy used by striker. Because of the
Docker components in the striker hosting stack, we configure envoy to
connect to backends on 127.0.0.1 rather than the default FQDN so that we
can avoid IPv6 usage (which is not well supported by Docker even in host
network mode).

Bug: T306469
Change-Id: I22397754468abe1de3fed12a3e7e1fdff8d6d336