labweb: point tlsproxy envoy at %{facts.ipaddress}:8080 for striker
Add explicit hiera config for the local tlsproxy to point striker
requests to our Docker container as an upstream.

Because of the Docker components in the striker hosting stack, we
configure envoy to connect to backends on the local primary IPv4 address
rather than the default FQDN so that we can avoid IPv6 usage (which is
not well supported by Docker even in host network mode). The 127.0.0.1
address is not used because of an explicit 127.0.0.1:80 vhost which
takes precedence over selection of named vhosts bound to *:80 (T113090).

Bug: T306469
Change-Id: Ie47292868dfdcb735493e01d8c3c0ac09ab0868b