gitlab: switch gitlab2001 back to "insetup" role
gitlab2001 has been decom'ed partially but not fully
(due to bugs with the cookbook).

This means right now though it still has the puppet role
on it but puppet can't run because the host has already been
removed from the list of allowed_hosts for acme_chief, to request
certs.

To avoid having a host with broken puppet but also allow the cookbook
to run normally, remove the role but don't remove the host from site.pp
entirely.

Instead set it back to "insetup". This way it has a role and cookbook
should work while we also avoid broken puppet alerts until then.

Bug: T307142
Change-Id: Idc0a38e910447d54f5b184ae4fc045a440988d61