cloudelastic: Increase primary cluster heap from 45G to 55G
cloudelastic has been struggling lately. In review it appears that
growth in the term dictionaries may have grown beyond what the
current heap can support. There is some hope that elastic 7.10 will
resolve some of this (they made some improvements to memory usage
around 7.7), but for now we need the clusters to have enough memory
to not be eternally stuck in Old GC Hell.

Bug: T309648
Change-Id: I325e3e8eccf7ec1cd6aeffed9e380fb96bd76d12