gitlab: fix IPs, hostname and regex for blackbox check
This change fixes the IP addresses and hostnames used for
backbox checking. By default prometheus is using the
primary interface, which is not serving 80/443. Furthermore
the check uses the actual service/hostname of the GitLab
instance.
The regex is changed from "DevOps" (which is not present
in the http body) to the headline which is served by the
/explore page ("Discover projects, groups and snippets").
Additionally, alerts for the non-production instance
(not active_host) will not be critical.

Change-Id: Ia3f103b0c97be7396c1c5239418a74edf8e323e1