gitlab_runner: Handle changes to runner config
The previous implementation used a GitLab config template to puppetize
most config variables. However, config templates are only used during
registration and subsequent changes to them are never re-loaded without
a full (de/re)-register cycle which results in other unwanted side
effects (see https://phabricator.wikimedia.org/T311241#8051974).

Managing the config file contents fully with puppet is not possible
because the registration process expects to be able to change the file,
writing its auth token back to it.

This change uses a workaround to support subsequent changes to the
config so that an already running runner service can reload it while
avoiding the race condition between puppet and the registration process.

 1) have `gitlab-runner register` write to a separate config file
 2) extract the auth token from file
 3) write our own temporary config that contains a placeholder for the
    auth token value
 4) use `envsubst` to substitute the placeholder with the auth token
    value and write the final config which the runner service should
    automatically reload

Bug: T311746
Change-Id: I5e60f33ea97e60b71d1b4ae385a08a35c1dc5e3b