Keystone: add a no-op userid hash generator
Keystone comes stock with only one default hash from
usernames to IDs. This has is invoked for any users
that are referred to outside of default domain.

We are about to make the ldap domain (named, confusingly, 'Default')
to a custom-config domain so that we can use mysql for all other
domains. In order to still have the lookup work in the no-longer-
quite-as-primary ldap domain, we need to skip the username->userid
hash so we can still use the existing userids in ldap.

Change-Id: I184894edcfb261eec7dc0d53b8d89a20934167c5