Keystone: rearrange how service domains are configured.
Previously we were trying to enumerate each service domain
with its own config. That turns out to not work at all;
we can only config ONE sql-based domain via config file.

To work around that, we're making sql the default driver,
and the old domain (called 'Default') the custom one
with ldap.

This will support one domain (everything that exists to date)
using ldap and all runtime-created domains backed with sql.

(And, for a refresh on context, we need service domains
backed with sql so that heat and magnum and create/delete
projects and users in these domains.)

Change-Id: Ib7839867274839b2f9a5b1aa0e34e0067a212cd4