haproxy: properly track client concurrency, & more
As it turns out, the previous implementation of misspassrate
would track the concurrency of cache miss/pass responses between
the time they were received from Varnish and the time they were
flushed to the user.  (Okay, I'm not 100% sure of that, but that's
certainly what it looks like and seems internally consistent.)

So instead, begin tracking concurrency as soon as the request is
received, and rename the table to just 'httpreqrate'.

Double both of the proposed thresholds, as the new interval is a strict
superset of the previous interval.

Drop 'newconnrate' as it hasn't been very useful.

Also log high client concurrency regardless of whether or not we'd drop,
for temporary investigation.

Hosts: cp2027.codfw.wmnet
Bug: T306580
Change-Id: I98785b547fd6c61e8be26519dd1b78c94042b6e9