gitlab: for now, only monitor the active host, not the replica
Our new prometheus::blackbox::http-based monitoring works great
except we are getting one, technically legit but also expected,
alert per day.

That is because we have the setup with the automatic backup on the
active host and restore on the replica.

During the time that "bash /srv/gitlab-backup/gitlab-restore.sh"
is running on the replica, https://gitlab-replica.wikimedia.org
will be down. It will recover once the restore ends succesfully.

Today it was 15 minutes from ProbeDown firing to resolved.

A semi-nice side effect is to see that the backup and restore
actually happens and works.. but of course we don't want a daily
false positive.

So first I thought we need to run the downtime cookbook, which
talks to both Icinga and alertmanager, somehow started by the
systemd timer, but .. on the cumin hosts?

Ideally we wouldn't have to know how long the downtime needs to be
but could send one start silence and one stop silence event.
I am not sure yet if that is possible.

For the short term, recognizing inspiration week, to avoid those
daily alerts it seemed an easy fix to skip the monitoring on the
replica and just monitor the active server.

Bug: T275170
Change-Id: I5ef9806dcb0ebe0f29470963b23222f00b748ada