opensearch: remove icinga::monitor::elasticsearch::old_jvm_gc_checks
For Cirrus this alert is covered by CirrusSearchHighOldGCFrequency, for
Opensearch I don't seem to remeber it ever firing.

If this shows up as a problem the CirrusSearchHighOldGCFrequency alert
will still trigger (though routed to search-platform for now), either
way we'll know sth is wrong, and in that case we can add a more specific
alert for Opensearch.

Bug: T288622
Change-Id: I3cb6b823537cebef3612cb3c00a203a7d284e3fa