Puppetize spark3 installation and configs using conda-analytics env
spark-env.sh will 'do the right thing' when working with
user python (conda) environments.  It will also always
default to using the conda-analytics python.

The system python should never be used as a default.

Outstanding TODOs:

- Automate building and upload spark assembly.
  https://phabricator.wikimedia.org/T310578

- Remove /usr/lib/airflow env conditional in spark3-env.sh.erb

- Automate installation of spark shuffle jar in Hadoop

Those changes are made to be applied only on the test cluster first to
perform some tests.

Bug: T312882
Bug: T295072
Change-Id: Iad71d4e422150cf03fed578c9a56e4f85c0dd4b6