Add roles and cumin aliases for the new dse_k8s cluster
We now have hosts the hosts installed for running the DSE Kubernetes
cluster, so we can beging to bootstrap it. This commit adds roles for
these hosts as well as cumin aliases for them.

At present the majority of the included profiles are commented out from
the role files. The idea being that the final state will be to have all
of these uncommented once the bootstrapping is complete. These profiles
will not work without additional hiera changes and secrets.

Bug: T310170
Bug: T310171
Change-Id: Ia6ff945ac478e3c85e172723c524a564943e640d