ci: enable docker on machine start
contint2001.wikimedia.org got rebooted but Docker did not start
automatically. The systemd docker.service unit is marked as disabled and
has to be enabled explicitly.

The trick is to have Puppet to explicitly enable the service:

    systemctl enable docker.service

The reason, found by Moritz, is T300682 migrated us from the Debian
maintained package docker-io package to docker-ce which misses the units
activations.

Bug: T313119
Change-Id: I702d6bc0fa66929f1ab4aba4b4f232cc42024b6e