alertmanager: switch IRC channel for gitlab (serviceops-collab) alerts
per serviceops meeting discussion:

These "receivers" we have created for the serviceops-collab subteam
and their services, such as gitlab and OTRS, are currently alerting
on the #wikimedia-serviceops channel but this is undesired.

First, move their output to #wikimedia-serviceops-test wait a few days
and eventually move it to #wikimedia-operations if it's stable.

Alternatives to this would be to drop the IRC alerting part completely
but keep mail and automatic task creation, or using yet another channel.

We do not want to unify receivers for serviceops-core and serviceops-collab
because we do already have separate Phabricator tags for tickets.

Personally, I would not recommend the "yet another channel" option and just
used -serviceops for initial testing/experimenting with the blackbox
alerts and notification types.

Change-Id: I8a5f412e0ec6b2e9214b0d61b3d7d988f950dcc9