durum: add IP version (v4/v6) in test result
Currently, durum does not display to the user if they are connected over
IPv4 or IPv6; this information might be helpful for the user and also
for debugging. This commit returns the IP version in the JSON result,
which we will then display on the durum frontend.

This is achieved by separating the IPv6 blocks and returning a JSON
response based on that.

Hosts: O:durum
Change-Id: I499425c24726ca63f2e3704cb0853db46c60afb8