nova: properly configure the number of api and metadata-api workers
There's a setting in nova.conf for this but it's ignored in favor of the
uwsgi config. This patch duplicates the hiera-dictated config into the
uwsgi.ini files.

This patch also consolidates the compute_workers and metadata_workers variables
as they've always been set to the same value and are likely to need adjusting in tandem
if at all.

Change-Id: I1b2df7e3f4ed2edf8ebdcb082e1a1dcd69b835ac