apifeatureusage: Temporarily remove index template during 6->7 transition
Elasticsearch 6.x and 7.x need slightly different index templates, but
for a week or two we intend to have this same bit of code talking to
an es6 cluster in one dc and an es7 cluster in the other dc. Prior to
upgrading the first cluster we need to remove the template from the
logstash config so it doesn't complain about failing index template
updates, then once the transition is complete we need to configure
it with the updated elastic 7 template.

Removing this template definition from the logstash output definition
should not remove the template from the elasticsearch clusters and
indices will continue to be created. Testing shows that es7 will
continue to work and create indices with the es6 template as long
as the template is using _doc as the mapping type.

Bug: T313434
Change-Id: I265a4da440221578c083c35b5ec28e82e2d067fa