gerrit: add hiera settings and IP for new replica gerrit2002
The goal is to add the gerrit puppet role on new host gerrit2002
but without causing any conflicts, especially regarding to the
additional service IP this adds to the network interface and
everything else related to replicating and syncing app data.

The regular setup is just one master and one replica but temporarily
we are setting up a second replica to then replace the old one.

Bug: T313250
Bug: T243027
Bug: T313972
Change-Id: Ib65aa6cf58e81251f17c37326ecc4e45f9c8c240