gerrit: add gerrit2002 to firewall rules for cluster support
This common list of "all gerrit servers" is used in these firewall
rules:

    # ssh between gerrit servers for cluster support
    $gerrit_servers_ferm=join($gerrit_servers, ' ')
    ferm::service { 'gerrit_ssh_cluster':
        port   => '22',
        proto  => 'tcp',
        srange => "(@resolve((${gerrit_servers_ferm})) @resolve((${gerrit_servers_ferm}), AAAA))",
    }

So it allows ssh to port 22, v4 and v6, between gerrit servers.

Keeping in mind that each gerrit server has 2 IPs, this is the
server and not the service IP.

As the comment says it is for "cluster support".

Should be safe to add gerrit2002 here before we add the gerrit role on
it unless I am missing something else where this is used.

Bug: T313250
Bug: T243027
Change-Id: I0a3a201269e7b923eddc40b1e65970ce6c37cfc3