gerrit: add gerrit2002 to puppetized known_hosts file
We do have a puppetized known_hosts file on gerrit servers,
should be for cluster support / ssh between gerrit servers.

gerrit2001 is in there. New host gerrit2002 is not.

Looks like we have to add it.

This is from /etc/ssh/ssh_host_rsa_key.pub.

Interestingly what is here in the repo and what is on the
servers matches for gerrit2001 but not for gerrit1001.

Bug: T313250
Bug: T243027
Change-Id: Ifa26ad9480af5a7b5fe1b65e82f65b025a4d41d2