jwt_authorizer: Provide microservice for JSON Web Token authorization
Defined jwt_authorizer module which includes package installation and a
`jwt_authorizer::service` resource for starting up distinct instances of
a jwt-authorizer daemon and token validation template. This service can
be used to validate JSON Web Token bearer tokens and perform any number
of assertions on token claim data.

The originating use case for this service is to provide integrated auth
between our GitLab instance and Docker registry, but it is a general
solution for JWT auth integrations.

Bug: T308501
Change-Id: I1e31f7ce32432d4e8c1c5cab83a7da3fb02d051a