admin: add Aline Bruenger to ldap_only admins (wmde,nda)
Aline Bruenger is a WMDE employee. Like other WMDE employees
this is for the 'wmde' and 'nda' LDAP groups.

The 'wmde' group gives certain Gerrit access in WMDE repos.

The 'nda' group gives logins at the various non-public
web UIs, very similar to 'wmf'.

NDA has been signed at T312220#8070321.

When this is merged there is still a manual step needed to
actually add the user to the 2 groups.

Bug: T312220
Change-Id: I10285b7260d25dc8f980bb90234f14c676a98dbb