C:ssh::client: Drop select from known_hosts file
Currently we do an inverse select aka reject for host aliases starting
with '---'.  Looking at the data in the database there are no aliases
that would be rejected by this as such this change should be a no op.

I also checked with joe who couldn't remember why this was originaly added
possibly some bug that's been fixed in the mean time, or some
functionality we never ended up using.  Either way it seems safe to go

Hosts: sretest1001.eqiad.wmnet
Change-Id: Id8fd9af564daa7991a8936c0739574a924899322