swift: stop flinging thumbnails at other DC in rewrite.py
In theory, the eventlets spawned to copy thumbnails to the other DC
are lightweight, but in practice we have been seeing problems in the
frontends sometimes seemingly related to timeouts from these threads.

Thumbnail generation should be active/active in any case now (cf
https://phabricator.wikimedia.org/T313102#8093002 ), so disabling
this.

Also remove inactivedc_request and inactivedc_thumborhost since these
are now unused in this code.

Bug: T313102
Change-Id: Ie3d691d7333b8147ba58227eb0a82e279d396b2f