Revert "services_proxy: Allow having both v4 and v6 AF enabled"
This reverts commit c736b1ce205abd952a7fca2af645e04fbaf97c61.

Reason for revert: build-envoy-config that consumes the files generated by this treats them as YAML instead of just concatenating them (as I thought). Those yaml files have 2 dicts with the same key in them and the parser will silently override the latter with the former by order of appearance in the file. That means we need some extra work before merging this again. Reverting for now to get back to the previously known state.

Change-Id: I73ebdff816b1565fbdcde0aa058497a6fe90157e