phabricator: add phabricator-roots on new phabricator hardware
Allows phabricator-root members to ssh to new phab servers
that are a simple hardware replacement, but before the
entire phabricator puppet role is applied.

It would happen automatically when the role is applied
so this is not giving any new access. It's just existing
group on new hardware that has the same role as old hardware.

Bug: T280597
Bug: T313360
Change-Id: I4584df1d2364d0ca8b2811bda8ba814e41fd6c3b