gerrit: turn gerrit2002 into a gerrit migration dest host
We have an existing role and profile for Gerrit migrations
which allows for setting up rsync etc before applying the
full gerrit role which includes an additional IP.

This should allow syncing data over to new host gerrit2002 during
provisioning.

It should also allow deploying gerrit to it and if not I will
add it here.

Bug: T313250
Bug: T243027
Change-Id: I1e2fa4bb62309e0d8bfecd0568f7b5f7c3821d1d