admin: add gerrit access groups to gerrit migration role
These are the same access groups applied on the production
role. There is no change involved who has what access.

It's just that we need a intermittent migration role where
we can start syncing data and deploying before switching to
new gerrit host.

And for that people in releng taking part in the current
"Decom Mission" sprint can take a look at it as well.

Bug: T280597
Bug: T313360
Change-Id: I9895d59bdeda48ce58dad7acfba32530f9aa9538