statistics::wmde::graphite: add max_runtime_seconds
These jobs run multiple scripts in parallel, which we want to run
periodically; if one of the scripts unexpectedly continues running for a
very long time, we would rather kill it than have it prevent the next
scheduled run (systemd won’t activate the timer’s associated unit again
if it’s still active).

For the daily timers, the specific duration of 20 hours is copied from
the update-openstack-mirror timer of profile::mirrors::openstack, but
any other number between, say, 16 and 23 hours would probably work too.
For the minutely and weekly timer, go with a maximum execution time
somewhat below the period as well.

Bug: T314130
Change-Id: I9def3dc512463659744334c6648408b7893ecbb1