trafficserver: 9.x upgrade: add compatibility for session_sharing.match
In ATS 8.x, proxy.config.http.server_session_sharing.match if used in
the configuration remap plugin assumes an INT value, which was set to 2
("ip"). In ATS 9.x, the INT value seems to have been deprecated in
favour of a STRING instead.

This mismatch resulted in a drop in the number of reused connections on
hosts using ATS9 (such as cp4032); setting server_session_sharing.match
to "ip" resolves that issue.

Note that this currently involves duplicating the params in
mapping_rules but that's fine for now as when ATS9 is rolled out fully
in production, we will just have one params key.

Bug: T309651
Hosts: cp4032.ulsfo.wmnet, cp2038.codfw.wmnet
Change-Id: I5056da9bd7a1a3e7ef43e0d85f2dd5dee844b110