phabricator: use migration role for pre-syncing data
Just like for Gerrit, add a role for setting up rsync
and requirements to let us sync user data from old
server to new server.

Also switch the exim to non-standard setup already
as this will be applied with the full phabricator role
but makes the diff a lot smaller and easier to review.

And handle shell access in the role avoiding hostname
hiera files.

Bug: T313360
Bug: T280597
Change-Id: Ice90572b8194b27b22513cfe1e753a1691675019