P:adduser: apply adduser before any packages are installed
We are still suffering race conditions where software can be installed
before we have correctly configured the system with the right apt
sources and the correct adduser config, As such create a dependency

This performs the desired action however we also need to be aware that
we don't also trigger puppetdb slowness[1] again.  I think this should be
fine as its a one to many relationship as opposed to a many to many
relationship (which is what we had in the previous issue) however we
should monitor puppetdb if/after this CR is merged

Its worth noting i did try this CR using stages (see previous PS's)
but that got me into some dependency cycle hell but happy to consider
that over this.  I think both options have there down sides, however i
feel we need something in place to ensure we get a more reproducible
build

Is also worth mentioning i have tested this change with beaker and after
its applied we still achieve a clean build from scratch.

[1]https://phabricator.wikimedia.org/T263578#7322135

Hosts: sretest1002.eqiad.wmnet
Bug: T235067
Bug: T263578
Change-Id: I262554a7edbfc23d1a8616dce439d6297135bf34