Add an option to use the PKI for etcd intra-cluster certificates
We would like to be able to use the CFSSL based PKI and the 'discovery'
intermediate CA to issue certificates for use with etcd clusters.
However we do not wish to change the configuration for existing etcd
clusters.

Therefore a new boolean value 'use_pki_certs' has been added to the
etcd::v3 profile, which enables this functionality. It defaults to
false.

We are currently bootstrapping a new etcd cluster for the dse-k8s
project, which will be the first cluster to use the PKI for its
certificates.

Hosts: P:etcd::v3
Bug: T313129
Change-Id: I04aa7729e52f865b0870fe7c97d9e5bc54fae4cb