Bootstrap etcd on the dse_k8s_etcd cluster
We are bootstrapping a new etcd cluster in support of the dse_k8s
kubernetes cluster. The hosts and roles exist already, but this commit
addds the configuration required to bootstrap the etcd cluster.

DNS SRV records already exist for this new cluster. The certificates
will be automatically generated due to the use of the `use_pki_certs`
parameter.

In addition we are defining two new clusters in the wikimedia_clusters
hiera structure, which will be used to refer to the dse_k8s and
dse_k8s_etcd clusters.

Hosts: O:etcd::v3::dse_k8s_etcd
Bug: T313129
Change-Id: Ieaef9761fba925c173322c8c030b7df6b97beca5