acme_chief: add gitlab-replica-old to certificate SANs
Arnold did a failover of a replica which worked
but also reported a certificate error with gitlab-replica-old.

This is the list of hosts that should show up on the certificate
created Letsencrypt that we get via our own acme_chief* hosts.

That's why it's added here and not in the gitlab code itself.

Change-Id: I09860fccf8672c47c9026d28921bbe36f4a68edd