icinga: Fix privacy policy check for the mobile wiki
Apparently, the html for the mobile wiki has changed and
simplified, based on my research with Internet Archive.

The HTML for the privacy policy on mobile now is the same
as with the desktop site.

Keeping the checks, however, separate because it could
diverge again in the future.

This should fix the ongoing alert:
en.m.wikipedia.org - Ensure legal html en.m.wp

Change-Id: I84111c94ff67e0e1bd8c5b1f1719f6e933700f28