nova.conf: remove auth_strategy = keystone
This produces the following deprecation warning:

Deprecated: Option "auth_strategy" from group "api" is deprecated for removal
(#012The only non-default choice, ``noauth2``, is for internal development and#012testing
purposes only and should not be used in deployments. This option and#012its middleware,
NoAuthMiddleware[V2_18], will be removed in a future release.#012).  Its value may be silently
ignored in the future.

Change-Id: I15a179a6cf369ae9e9f572daeb78d84885c225e9