Change ownership of etcd certificates for cfssl usecase
When using the default option of the get_cert PKI method, the
certificates were generated with ownership limited to root. This meant
that the etcd user was unable to read the certificates and start the
etcd server.

This change ensures that those certificates are owned by the etcd user.

Note that resource ordering may present an issue when bootstrapping a
cluster, since the etcd user is referenced before being created by the
package.

Bug: T313129
Change-Id: I86416a3cee9f20166de1e525fed21dc9cdc20294