Change the output directory that cfssl uses for etcd
We previously created the certificates in the default /etc/cfssl/ssl
directory but the etcd user did not have the necessary rights to read
this directory.

This change causes the certficiates to be created in /var/lib/etcd/ssl
instead and changes the owner of that directory to be etcd as well.

Bug: T313129
Change-Id: Ib73f0e971443e131b09549578722cb8f7075b9d5