Add additional network device info to puppet facts
This adds to the existing primary_interface.rb custom facts code,
adding additional information where possible about Vlan and Bridge
devices.  The goal is to export this so we can parse the data
in the netbox import script, and model these relationships correctly
in Netbox.

It adds up to 4 new vaiables to the elements under
networking.interfaces:

 kind:          netdev 'info_kind' reported, i.e. 'bridge' or 'vlan'
 dot1q:         802.1q tag configured if interface kind is vlan
 parent_link:   parent interface if interface kind is vlan
 parent_bridge: parent bridge device if int is bound to one

Example of the augmented networking vars can be seen here:

https://phabricator.wikimedia.org/P32328

The code depends on the '--json' output of iproute2.  This does
not seem to be supported for stretch hosts, so I've wrapped the code
adding the new data in a check on the os release.

This is my first time doing any Ruby code so it's probably an
absolute nightmare. Sorry!

Bug: T296832
Change-Id: I142e27f9eac91a3fedafbedc5a9b6022e682187b