phabricator: make it possible to globally enable/disable vcs setup
Make it possible to globally enable/disable the entire Phabricator
VCS setup.

And do it _before_ we apply the prod role to new machines.

This way we can be 100% sure we don't cause issues like assigning
the same IP to multiple hosts during migration and we also don't
have to worry about cleaning up / absenting resources and sort
out later what we don't need anymore after git-ssh.wikimedia.org
is shut down.

Just do not not install any of this on the new hosts and see if anything
is missing instead.

Also keep it noop on existing prod hosts at the same time.

Bug: T280597
Change-Id: Ic2ec4e64d9bb1f831117027451805b67bee8cbe9