C:ferm: update ferm to use restart-or-reload instead of restart
The firm service resource abuses puppet a bit in that we need to
change the behaviour of the service status.  normally a failed
status command indicates to puppet that a service is stopped.  however
we use it to ensure that the rule base on disk matches the rule base
loaded by the kernal.  if not we want to reload ferm

Currently this works by calling systemd restart.  however that first
stops and then starts ferm which means that if we have a syntax error we
end up unloading the rule base and then are unable to load the new
rules, leaving the host without any firewall rules.

This PS updates the service resource to instead use reload-or-restart.
This means that if the services is already started we will attempt a
reload, which in the case of a syntax error will cause a failure to the
systemd unit but will not change the rule base i.e. the last known good
rule base will stay in place.

We use reload-or-restart over just reload so that we can also ensure we
start the service if it is stopped

Bug: T315305
Hosts: sretest1001.eqiad.wmnet
Change-Id: Ibfb97c2169ce3d875cbe10153bdb1fa05fa57451