phabricator: move lvs::realserver inclusion to profile, create use_lvs parameter
The inclusion of lvs::realserver is here for the git-ssh pybal service.
It includes things like adding a service IP to the loopback interface
for a service behind LVS.

The enable_vcs parameter was introduced in Ic2ec4e64d9bb1f831 to make
it possible to disable the entire VCS/git-ssh setup.

We set this parameter to false on new replacement servers because
we don't even want to install any of this to avoid issues such as
T224677 or T172478. It was the point of T296022 to shut this down
before migrating.

Now if we apply the production puppet role to new machines
we still get the LVS setup so move this out of the role and into
a profile so that we can make it depend on $enable_vcs.

Also instantiate the class instead of including it because the style
check won't like that afair.

After initial review add another new parameter use_lvs to toggle
inclusiong of the LVS class.

Should stay noop on prod phab servers but allow us to add the
role on new servers.

Bug: T280597
Bug: T296022
Change-Id: I0abe6d238d065d59b45bbc31eab21a1626d4b929