phabricator::migration: add phd with systemd::sysuser, reserve UID 920
Add the phd user in the Phabricator migration role so that when
we rsync data over we can already let phd own it.

But use the newer systemd::sysuser we are supposed to use
now to create system users, which the main Phabricator class doesn't
do yet.

We do want this to reserve our phd UID properly and avoid future
issues with data syncing.

per https://wikitech.wikimedia.org/wiki/UID

Reserve UID 920 for phd on newer servers after making this edit:

https://wikitech.wikimedia.org/w/index.php?title=UID&type=revision&diff=2004642&oldid=2004195

Bug: T313360
Change-Id: I924be1dc253b3466c144ea8736acfe7f95b9c217