phabricator: replace user{} with systemd::sysuser, only on new hosts
We should replace system user creations
with systemd::sysuser and want to make sure we always
have the same uid on all machines using the Phabricator role

Plus the ID 498 we have been using previously should be replaced by anew UID
in the range above 900, it's been defined as 920.

Add a temp. parameter to switch it only on new machines.

We will use the correct and new UID 920 there and later
remove all of this once old servers have been shut down.

Bug: T313360
Change-Id: Ia6eff9dde52e5790e70cf2628f969cc87502aa39