node_pinger: use jumbo frames
This will help also detect network configuration issues.

Makes sure that jumbo frames are only sent between osds, and osds and
cloudvirts.

It also adds two extra labels to the metrics generated, the ip that
was pinged, and if the ping was a jumbo frame or not.

Note that we don't yet have jumbos from cloudvirts<->osds (T330075).

Bug: T314870
Bug: T329711
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Iae5735e13e4b757c1046a74d1e3136198d80467b