p:ceph::osd: get the os disks by size
As the device names might change between reboots, there's no certainty
anymore that sda and sdb would be the one for the OS, or that if they
are, it will stay like that after a kernel upgrade.
This patch uses the disk sizes with the assumption that os disks will
be smaller than the ones used for the OSD daemons.

Bug: T314870
Change-Id: Ib1ec8d6a4d988dabe2eb68327836344df1e40348
Signed-off-by: David Caro <dcaro@wikimedia.org>