doc: properly redirect back compat URLs
Our rewrite rules prevent `Redirect` from working they are entirely
ignored which broke the few back compat redirect we had.

I noticed it after 2d05626611 got deployed to redirect /mw-tools-scap/
to /scap/ which still served the old content. The other redirects for
CollaborationKit and Wikibase issued a 404 as well.

The reason is `RewriteRule` directives run first regardless of the order
of appearance in the configuration file which is hinted at in:
https://httpd.apache.org/docs/2.4/rewrite/avoid.html#redirect

Convert the mod_alias Redirect directives to mod_rewrite RewriteRule.
Flag sets are:
- L: no further rule processed
- R: redirect issueing a 302 (the default)

There is no need for QSA since the substitution pattern does not have
parameters.

I have added a httpbb code covering the path redirection.

Bug: T315541
Change-Id: I462708a177ded0d0439429ff47662c454783342f