swift: ms-be2067/sdc1 has failed
This drive has failed, and the server needs a new back plane (it fails
to recognise a replacement drive in slot 0); so mark this drive as
failed. Once the resulting rebalance has completed the ms-codfw
cluster should again be healthy enough to go back to the on-going
load/drain of servers.

Bug: T314049
Change-Id: I9e248955026f36c6b7e24b6d6a97b4bd0e72f722