phabricator: fix sshd listen address for phab codfw
Phabricator hosts used to have 2 separate ssh daemons.

One for git-ssh and one regular one for admins.

Both listen on their own IP.

On new phabricator hosts we don't want to support git-ssh
anymore, so we removed the setup for it (LVS, service IP on loopback,
etc).

But then there was also this parameter set to change the listen
address for the standard sshd server, not the one installed by
Phabricator but the default one all machines get from puppet.

In eqiad it is applied via ./hosts/phab1001.yaml but in codfw
it was applied via role/codfw.

This meant once we have more than one host in codfw, phab2001
and phab2002 right now, puppet tries to make the sshd on phab2002
listen on the IP of phab2001 and that ... fails in a way that
actually takes down sshd.

Bug: T280597
Bug: T313360
Change-Id: Ib5f721c85c214c7d9d5ac1889c4a10e69efcab43