phabricator: move vcs and LVS settings from common to phab2001
These settings should only be needed on old Phabricator hosts
but not anymore on new Phabricator hosts, after we decom the
git-ssh service behind LVS.

So move them out of the common file to the one specific to phab2001.

Then once phab2001 is gone it will be an automatic cleanup with nothing
left behind.

Bug: T280597
Bug: T296022
Change-Id: I4765a4d6e48d8733aaa035ceca342744edd373d7