site: add phabricator role on phab1004
After we did phab2002 it should now also be possible in eqiad.

But we need to be careful here since it's the prod DC and before
we had things like duplicate IP.

Previously in edcd5c8ae71 disabled phd service, dumps and logmail.

Do use systemd::sysuser to create the sysuser and try again
if there is a previously discussed issue when doing this
on a new host for the first time and more than one system
user is created.

Bug: T280597
Change-Id: I503f9c9f14c246b46bdd019095180c25f3649971