dumps/phabricator: switch phab dumps host from phab1001 to phab1004
dumps are created by talking to the mysql-slave, regardless
of the Phabricator server being the "active" one or not.

So we can switch over before we switch the main app from
old to new hardware.

I570768bb03df32f made it so that this is now only a Hiera edit
and not directly touching dumps classes.

I9a30c6824fc9f5 enabled dumping on phab1004.

This change switches where clouddumps fetch the public phab
dump from.

There is still room for improvement here because we change
the same name in multiple places, should be just one?

Bug: T280597
Change-Id: I58ed2bdc6725b7caa1da96c411d616065457f648