Run the initsitestats period job on a daily basis
After new wikis are created and their content is imported by Incubator,
it currently takes up to two weeks for the statistics to be updated on
Special:Statistics.

Running the update script does not seem to be very expensive, so it
seems appropriate to run the script daily to fix potential drifts
instead of letting them persist for up to two weeks.

Bug: T315121
Change-Id: Ie98bf696620a3c8a73efac2509dbc9af4f062a62