Allow cloud_provider_enabled
In order to use a version of kubernetes beyond 1.18 we need
to pull a copy of hyperkube that is from an external source.
This needs to be explicitly allowed by passing the template
cloud_provider_enabled=true at creation. However this is not
allowed unless cluster_user_trust=true. This change will
allow for that.

https://docs.openstack.org/magnum/latest/user/

Bug: T280792
Change-Id: I0b4f93d1a8a767cf28990d43ac733f3bb1d40f41