query_service: Avoid passing content body to internal auth endpoints
When requests are POST'd to the commons query service they stall for
30 seconds and then timeout. On review this looks to happen when nginx
passes the Content-Length header to the auth endpoint without any
content and jetty, hosting the oauth endpoints, waits for the body
content that never arrives.

It's not clear why 'proxy_pass_request_headers off` on it's own isn't
sufficient, but in testing we need to explicitly blank the
Content-Length header as well. This was tested in the mw-oauth-proxy
integration environment, in I4657ba699c.

Bug: T306899
Change-Id: I5966f0306967e8aebe0c73613d2172b82c9b9bef