phabricator: move LVS IPs for git-ssh service from role/eqiad to phab1001
The LVS service IPs for the git-ssh service should move from role/eqiad
to the host level to be applied only on phab1001.

We don't want to support this service on new hosts and are during a
migration and need to avoid any complications from having this applied
on a role/eqiad level since there are temporarily 2 servers in eqiad.

Once phab1001 is eventually removed this will naturally disappear.

Also the other "LVS-enable" settings and VCS settings have already
moved.

Bug: T296022
Bug: T280597
Change-Id: Ic23e46e82493dc1f249a4a9897d2a98bbcc010ef