trafficserver: Hide non session cookies during cache lookup
Prevent non-session cookies from impacting on caching performance
by hiding them before cache lookup and restoring them afterwards.

misc is misterious so as a first approach we don't hide cookies
for cache lookup if it's a misc request according to varnish.

A restart of trafficserver is required to make this change effective
as it relies in two new global hooks being added to default.lua:
 * do_global_post_remap()
 * do_global_cache_lookup_complete()

Bug: T316338
Depends-On: Ibcdef27784a8ca8cd4cfce23d963605b3985d2d6
Change-Id: I44bbb04e0cdbbe6ad6ef539f552412fe2a25c2b9